Simulate Aggregate Expected Deaths

Simulate Aggregate Expected Deaths

Agesim_infxn_2_death(
  fatalitydata,
  infections,
  m_od = 14.26,
  s_od = 0.79,
  curr_day,
  spec,
  sens,
  demog,
  sero_delay_rate,
  simulate_seroreversion,
  sero_rev_shape = NULL,
  sero_rev_scale = NULL,
  smplfrac = 1,
  return_linelist = FALSE
)

Arguments

fatalitydata
dataframe; The strata-specific fatalities to simulate given a probability of infection and a noise effect. The column names: strata, ifr, rho, and Ne correspond to (patient) strata, infection-fatality ratio, the probability of infection (i.e. a probalistic attack rate), and a noise effect, respectively.

infections
integer vector; The infections for each day up to the current day.

m_od
double; The mean of the onset of infection to death (gamma distribution).

s_od
double; The coefficient of variation of the onset of infection to death (gamma distribution).

curr_day
numeric; Current day of epidemic (considered up to but not including this day).

spec
double; Specificity of the Seroprevalence Study

sens
double; Sensitivity of the Seroprevalence Study

demog
dataframe; Strata-specific population (demographic) counts. The columns names strata and popN correspond to (patient) strata and the number of individuals within that strata. The demography strata must match the fatalitydata strata. Only considered if simulate_seroprevalence = TRUE

sero_delay_rate
double; Rate of time from infection to seroconversion, assumed to be exponentially distributed

simulate_seroreversion
logical; Whether seroreversion (due to waning of antibodies) should be simulated or not

sero_rev_shape
double; The shape parameter of the Weibull seroreversion distribution

sero_rev_scale
double; The scale parameter of the Weibull seroreversion distribution

smplfrac
numeric; Sampling fraction for the observed seroprevalence study (assumed to be a simple random sample of all infected)

return_linelist
logical; Whether or not the linelist that was used to create the observed marginal data should be returned. N.B. the linelist can be quite large/burdensome for memory depending on population size and number of days considered.