Simulate Cumulative Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve

Simulate Cumulative Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve

ExpGrowth_CumulativeDeath_sim(
  casefat,
  I0,
  r,
  curr_day,
  m_od = 18.8,
  s_od = 0.45
)

Arguments

casefat
dataframe; column names age, cfr, and pa that correspond to age-band, case-fatality ratio, and the attack rate in given age band, respectively

I0
integer; number of infected individuals at the beginning of the epidemic

r
double; the expontential growth rate of the epidemic

curr_day
numeric; current day of epidemic

m_od
double; the mean of the onset of infection to death (gamma distribution)

s_od
double; the coefficient of variation of the onset of infection to death (gamma distribution)