Draw the Inferred Seropevalence Curves both Adjusted and Unadjusted for Specificity and Sensitivity with the Rogan-Gladen Estimator

Draw the Inferred Seropevalence Curves both Adjusted and Unadjusted for Specificity and Sensitivity with the Rogan-Gladen Estimator

draw_posterior_sero_curves(
  IFRmodel_inf,
  whichrung = "rung1",
  dwnsmpl,
  by_chain = TRUE
)

Arguments

IFRmodel_inf
R6 class; The result of the IFR Model MCMC run along with the model input.

whichrung
character; Specify which rung to sample from (default is rung1)

dwnsmpl
integer; Number of posterior results to draw -- weighted by posterior prob

by_chain
logical; Whether or not credible intervals should be reported with respect to individual chains (TRUE) or not.

Details

Given sampling iterations with posterior-log-likes greater than or equal to a specific threshold, posterior results for the linear spline are generated. Assumed that the spline was fit in "un-transformed" space