Simulate Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve

Simulate Expected Deaths for a Given Day Assuming Exponential Growth of the Incidence Curve

sim_infxn_2_death(
  casefat,
  I0,
  r,
  m_od = 18.8,
  s_od = 0.45,
  min_day = 1,
  curr_day,
  level,
  expgrowth,
  infections
)

Arguments

casefat
dataframe; column names age, cfr, and pa that correspond to age-band, case-fatality ratio, and the attack rate in given age band, respectively

I0
integer; number of infected individuals at the beginning of the epidemic for exponential growth

r
double; the expontential growth rate of the epidemic

m_od
double; the mean of the onset of infection to death (gamma distribution)

s_od
double; the coefficient of variation of the onset of infection to death (gamma distribution)

min_day
numeric; first day epidemic was observed

curr_day
numeric; current day of epidemic, considered up to but not including this day

level
character; Either "Cumulative" or "Time-Series" if expected deaths should be returned as a total or for each day, respectively

expgrowth
logical; Simulate exponential growth or not

infections
integer vector; The infections for each day