Wrap the Metropolic-Coupled MCMC Framework from Dr. Jacoby

Wrap the Metropolic-Coupled MCMC Framework from Dr. Jacoby

wrap_drjacoby_mcmc(
  data,
  df_params,
  misclist,
  LogLike,
  LogPrior,
  level,
  burnin = 1000,
  samples = 1000,
  chains = 3,
  rungs = 1,
  GTI_pow = 3,
  coupling_on = T,
  pb_markdown = F,
  silent = T
)

Arguments

data
a vector of data values. When using C++ versions of the likelihood and/or prior these values are treated internally as doubles, so while integer and boolean values can be used, keep in mind that these will be recast as doubles in the likelihood (i.e. TRUE = 1.0).

df_params
a dataframe of parameters. Must contain the following elements:

name - the parameter name.

min - the minimum value of the parameter. -Inf is allowed.

max - the maximum value of the parameter. Inf is allowed.

init - the initial value of the parameter.

misclist
list; Required list object that must contain the following items: elements:

min_day - the presumed minimum day of the epidemic (required only for time-series likelihoods)

curr_day - the presumed minimum day of the epidemic.

pa - A vector of the attack rates for the given age bands

LogLike
global function;

LogPrior
global function;

level
character; Must be either Time-Series or Cumulative

burnin
the number of burn-in iterations.

samples
the number of sampling iterations.

chains
the number of independent replicates of the MCMC chain to run. If a cluster object is defined then these chains are run in parallel, otherwise they are run in serial.

rungs
the number of temperature rungs used in Metropolis coupling (see coupling_on).

GTI_pow
the power used in the generalised thermodynamic integration method.

coupling_on
whether to implement Metropolis-coupling over temperature rungs.

pb_markdown
whether to run progress bars in markdown mode, meaning they are only updated when they reach 100 being printed to markdown files.

silent
whether to suppress all console output.